+++ software/firmware/badge/video_lld.c
@@ -264,7 +264,7 @@ videoPlay (char * path)
 	max = (ch->cur_vid_size + ch->cur_aud_size) + sizeof(CHUNK_HEADER);
 
 	btmp = chHeapAlloc (NULL, (max * 2) + CACHE_LINE_SIZE);
-	buf = (uint8_t *)roundup ((uint32_t)btmp, CACHE_LINE_SIZE);
+	buf = (uint8_t *)roundup ((uintptr_t)btmp, CACHE_LINE_SIZE);
 
 	p1 = buf;
 	p2 = buf + max;
